Fix come clipboard issues for tokens and consolidate clipboard functionality
